Add @socialincome/shared & firebase admin (firestore) integration
