+++ seed/auth_export/accounts.json
@@ -1,1 +1,1 @@
-{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"Lk6iS7tIuxtl4jZXYeu9DW3PX68Y","createdAt":"1666207539605","lastLoginAt":"1668257411691","displayName":"Admin","providerUserInfo":[{"providerId":"google.com","rawId":"7263616607754755158306200924200494132545","federatedId":"7263616607754755158306200924200494132545","displayName":"Admin","email":"admin@socialincome.org"}],"validSince":"1668257365","email":"admin@socialincome.org","emailVerified":true,"disabled":false,"lastRefreshAt":"2022-11-12T12:50:11.692Z"}]}
+{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"Lk6iS7tIuxtl4jZXYeu9DW3PX68Y","createdAt":"1666207539605","lastLoginAt":"1669470148332","displayName":"Admin","providerUserInfo":[{"providerId":"google.com","rawId":"7263616607754755158306200924200494132545","federatedId":"7263616607754755158306200924200494132545","displayName":"Admin","email":"admin@socialincome.org"}],"validSince":"1669470065","email":"admin@socialincome.org","emailVerified":true,"disabled":false,"lastRefreshAt":"2022-11-26T13:42:28.332Z"}]}
