+++ website/__tests__/pages/transpararency/finances/[currency].test.tsx
@@ -1,30 +1,9 @@
 import { render } from '@testing-library/react';
-import { NextURL } from 'next/dist/server/web/next-url';
-import { instance, when } from 'ts-mockito';
-import Finances, { financesMiddleware } from '../../../../pages/transparency/finances/[currency]';
-import { mockRequest } from '../../../__utils__/mocks';
+import Finances from '../../../../pages/transparency/finances/[currency]';
 
 describe('Finance page should', () => {
 	it('render page unchanged', async () => {
-		const { container } = render(<Finances currency={'CHF'} />);
+		const { container } = render(<Finances currency={'CHF'} balance={200} />);
 		expect(container).toMatchSnapshot();
-	});
-});
-
-describe('financesMiddleware should', () => {
-	it('return noop for unrelated path', async () => {
-		const mockedRequest = mockRequest();
-		when(mockedRequest.nextUrl).thenReturn(new NextURL('/', 'https://socialincome.org'));
-		const result = await financesMiddleware(instance(mockedRequest));
-		expect(result).toBeUndefined();
-	});
-
-	it('it returns redirect for /transparency/finances', async () => {
-		const mockedRequest = mockRequest();
-		when(mockedRequest.nextUrl).thenReturn(new NextURL('/transparency/finances', 'https://socialincome.org'));
-		// TODO mock different geolocations of users
-		const result = await financesMiddleware(instance(mockedRequest));
-		expect(result?.status).toBe(307); // redirect
-		expect(result?.headers.get('Location')).toBe('https://socialincome.org/transparency/finances/chf');
 	});
 });
