+++ website/package.json
@@ -17,6 +17,7 @@
 		"format-code": "prettier --write --config ../.prettierrc --ignore-path ../.prettierignore ."
 	},
 	"dependencies": {
+		"@socialincome/shared": "*",
 		"@types/node": "18.11.9",
 		"@types/react": "18.0.25",
 		"@types/react-dom": "18.0.9",
@@ -25,6 +26,7 @@
 		"firebase-admin": "^11.3.0",
 		"firebase-functions": "^4.1.0",
 		"next": "13.0.4",
+		"next-transpile-modules": "^10.0.0",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
 		"typescript": "4.9.3"
@@ -32,9 +34,9 @@
 	"devDependencies": {
 		"@testing-library/jest-dom": "^5.16.5",
 		"@testing-library/react": "^13.4.0",
+		"isomorphic-fetch": "^3.0.0",
 		"jest": "^29.3.1",
 		"jest-environment-jsdom": "^29.3.1",
-		"ts-mockito": "^2.6.1",
-		"isomorphic-fetch": "^3.0.0"
+		"ts-mockito": "^2.6.1"
 	}
 }
