+++ website/pages/transparency/finances/[currency].tsx
@@ -1,17 +1,20 @@
+import { doc } from '@socialincome/shared/src/firebase/firestoreAdmin';
+import { BankBalance, BANK_BALANCE_FIRESTORE_PATH } from '@socialincome/shared/src/types';
 import { GetStaticPaths, GetStaticProps } from 'next';
-import { NextRequest, NextResponse } from 'next/server';
 import Layout from '../../../components/layout';
 import { appConfig } from '../../../config';
 
 interface Props {
 	currency: string;
+	balance: number;
 }
 
-export default function Finances({ currency }: Props) {
+export default function Finances({ currency, balance }: Props) {
 	return (
 		<Layout title={'Transparency Page'}>
 			<section>
 				<p>You are seeing the transparency page for {currency}</p>
+				<p>Balance retrieved from firestore: {balance}</p>
 			</section>
 		</Layout>
 	);
@@ -22,10 +25,12 @@ export default function Finances({ currency }: Props) {
  */
 export const getStaticProps: GetStaticProps = async (context) => {
 	const currency = context.params?.currency as string;
-	// TODO import transparency stats from firestore
+	// TODO import proper transparency stats from firestore
+	const exampleFirestoreDoc = await doc<BankBalance>(BANK_BALANCE_FIRESTORE_PATH, 'main_1669470424').get();
 	return {
 		props: {
 			currency,
+			balance: exampleFirestoreDoc.data()?.balance,
 		},
 		revalidate: 60 * 2, // rebuild these pages every 2 minute on the server
 	};
@@ -43,19 +48,4 @@ export const getStaticPaths: GetStaticPaths = async () => {
 		})),
 		fallback: false,
 	};
-};
-
-/**
- * Forwards from /transparency/finances to /transparency/finances/[currency] according to the user's location.
- * This allows us to combine incremental static prebuilt pages and personalization based on user's location.
- * @param request
- */
-export const financesMiddleware = (request: NextRequest) => {
-	if (request.nextUrl.pathname.endsWith('/transparency/finances')) {
-		// TODO add logic leveraging the request.geo.country information
-		const redirectUrl = request.nextUrl.clone();
-		redirectUrl.pathname = redirectUrl.pathname + '/chf';
-		return NextResponse.redirect(redirectUrl);
-	}
-	return undefined;
 };
